Add Fizzbuzz tests and pagination tests, add page size selector and fix minor bugs

+++ priv/repo/migrations/20220921120013_create_favorites.exs
@@ -3,7 +3,7 @@ defmodule Fizzbuzz.Repo.Migrations.CreateFavorites do
 
   def change do
     create table(:favorites) do
-      add :number, :integer
+      add :number, :bigint
 
       timestamps()
     end
